+++ frontend/src/utils/useForm.tsx
@@ -12,7 +12,7 @@ export default function useForm(initial = {}) {
     setInputs(initial);
   }, [initialValues]);
 
-  function handleChange(e) {
+  function handleChange(e:any) {
     let { value, name, type } = e.target;
 
     // form inputs always save data as strings