+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.5] - 2024-02-26
+### Updated
+- Fixed build issues for frontend files
+- Fixed backend build issues
+
 ## [0.3.4] - 2024-02-23
 ### Added
 - Minor style updates